Moved Detail Pane Code Into Views
