+++ apps/zui/pages/detail.tsx
@@ -2,13 +2,13 @@ import {invoke} from "lodash"
 import React, {useEffect, useState} from "react"
 import {AppProvider} from "src/app/core/context"
 import AppWindowRouter from "src/app/router/app-window-router"
-import LogDetailsWindow from "src/js/components/LogDetailsWindow"
 import {Modals} from "src/js/components/Modals"
 import {Tooltip} from "src/components/tooltip"
 import initialize from "src/js/initializers/initialize"
 import TabHistories from "src/js/state/TabHistories"
 import Tabs from "src/js/state/Tabs"
 import {getPersistedWindowState} from "src/js/state/stores/get-persistable"
+import {DetailsWindow} from "src/views/details-window"
 
 export default function DetailPage() {
   const [app, setApp] = useState(null)
@@ -34,7 +34,7 @@ export default function DetailPage() {
   return (
     <AppProvider store={app.store} api={app.api}>
       <AppWindowRouter>
-        <LogDetailsWindow />
+        <DetailsWindow />
         <Modals />
         <Tooltip />
       </AppWindowRouter>
