+++ apps/zui/src/app/detail/NoSelection.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import {EmptyText} from "../features/right-pane/common"
+import {EmptyText} from "../../views/right-pane/common"
 
 const NoSelection = () => (
   <EmptyText>Select a value in the results to view details.</EmptyText>
