+++ apps/zui/src/app/routes/app-wrapper/app-wrapper.tsx
@@ -6,7 +6,7 @@ import {MainArea} from "./main-area"
 import {AppGrid} from "./app-grid"
 import useLakeId from "src/app/router/hooks/use-lake-id"
 import {DataDropzone} from "./data-dropzone"
-import RightPane from "src/app/features/right-pane"
+import RightPane from "src/views/right-pane"
 
 export default function AppWrapper({children}) {
   const lakeId = useLakeId()
