--- apps/zui/src/js/components/common/index.ts
@@ -1,5 +1,0 @@
-import HistoryButtons from "./HistoryButtons"
-
-export default {
-  HistoryButtons,
-}
