+++ apps/zui/src/views/columns-pane/columns-tree.tsx
@@ -4,7 +4,7 @@ import {ListItem} from "src/components/list-item"
 import {NodeRendererProps, Tree} from "react-arborist"
 import {FillFlexParent} from "src/components/fill-flex-parent"
 import classNames from "classnames"
-import {EmptyText} from "src/app/features/right-pane/common"
+import {EmptyText} from "src/views/right-pane/common"
 import {useSelector} from "react-redux"
 import Table from "src/js/state/Table"
 import {TableColumn} from "src/js/state/Table/selectors"
