+++ apps/zui/src/views/right-pane/common.tsx
