+++ apps/zui/src/views/right-pane/history/use-entry-menu.ts
