+++ apps/zui/src/views/right-pane/index.tsx
@@ -4,7 +4,7 @@ import DetailSection from "./detail-section"
 import styled from "styled-components"
 import {useDispatch} from "src/app/core/state"
 import {useSelector} from "react-redux"
-import Layout from "../../../js/state/Layout"
+import Layout from "src/js/state/Layout"
 import {DraggablePane} from "src/js/components/draggable-pane"
 import VersionsSection from "./versions-section"
 import AppErrorBoundary from "src/js/components/AppErrorBoundary"
