+++ apps/zui/src/views/right-pane/version-item.tsx
